# perf/perftool: Separate perf test and perftool_test