… perf/perftool.py