perftool test contains 2 subtests
     - perf test
     - perftool_testsuite
    
Separate these 2 into an independent tests for ease of use and maintenance.
[1/3] perf/perftool: Remove perftool test
[2/3] perf/perftool: Introduce perf test
[3/3] perf/perftool: Introduce perftool_testsuite

Signed-off-by: Sachin Sant <sachinp@linux.ibm.com>