@sacsant 2023-02-13
Before the fix against 6.2.0-rc7-next-20230208 using perf tool shipped with distro:

 (1/2) perf/perftool.py:Perftool.test_perf_test: STARTED
 (1/2) perf/perftool.py:Perftool.test_perf_test: FAIL: 5 Test failed (151.35 s)
 (2/2) perf/perftool.py:Perftool.test_perf_testsuite: STARTED
 (2/2) perf/perftool.py:Perftool.test_perf_testsuite: FAIL: 17 Test failed (282.55 s)

After the fix against 6.2.0-rc7-next-20230208 using upstream version of perf:
$ perf --version
perf version 6.2.0-rc8
$

The number of perf_test failures have reduced to 2 (from 5)
 (1/2) perf/perftool.py:Perftool.test_perf_test;run-run_type-upstream-41dd: STARTED
 (1/2) perf/perftool.py:Perftool.test_perf_test;run-run_type-upstream-41dd: FAIL: 2 Test failed (153.75 s)
 (2/2) perf/perftool.py:Perftool.test_perf_testsuite;run-run_type-upstream-41dd: STARTED
 (2/2) perf/perftool.py:Perftool.test_perf_testsuite;run-run_type-upstream-41dd: FAIL: 12 Test failed (223.10 s)