@disgoel 2023-02-14
@sacsant can you also add pre-requisite packages to run perf test otherwise few test skips.
clang, perf-debuginfo, kernel-debuginfo

one more request from dev was to run perf test after compiling llvm