@sacsant 2023-02-14
> @sacsant can you also add pre-requisite packages to run perf test otherwise few test skips. clang, perf-debuginfo, kernel-debuginfo
>

Is this for both distro as well as upstream as a target? I guess debug info packages are meant for distro only.
 
> one more request from dev was to run perf test after compiling llvm

same question, distro and/or upstream ??