@sacsant 2023-02-14
On second thought let me revisit this. The two test scenarios are separate and not testing same functionality. It will be better to split this into 2 separate tests -- one for perf test and another for perf-testsuite.

Will work on this and push new code shortly. Moving to draft stage.