@sacsant 2023-02-14
$ avocado run perf/perf_test.py 
...
 (1/1) perf/perf_test.py:Perftest.test_perf_test: STARTED
 (1/1) perf/perf_test.py:Perftest.test_perf_test: FAIL: 3 Test failed (51.76 s)
RESULTS    : PASS 0 | ERROR 0 | FAIL 1 | SKIP 0 | WARN 0 | INTERRUPT 0 | CANCEL 0
$
$ avocado run perf/perftool_test.py 
....
 (1/1) perf/perftool_test.py:Perftool.test_perf_testsuite: STARTED
 (1/1) perf/perftool_test.py:Perftool.test_perf_testsuite: FAIL: 13 Test failed (272.65 s)
RESULTS    : PASS 0 | ERROR 0 | FAIL 1 | SKIP 0 | WARN 0 | INTERRUPT 0 | CANCEL 0
$