@sacsant 2023-02-14
I will submit a separate PR to add upstream as a target support for perf_test.py