@sacsant 2023-02-21
> > I will submit a separate PR to add upstream as a target support for perf_test.py
> 
> @sacsant are you going to add pre-requisite packages in a separate PR ? If yes, this patch looks good.

Yes will address the pre-reqs along with upstream support in a separate PR.